BUG - Path should have no dependency
The path points to a directory which is not covered by the bulk-move plugin. The path relies on a different plugin (zenphoto-admin-chosen-dropdowns) which can't be found in the web (though it's mentioned on the Zenphoto website which is confusing: https://www.zenphoto.org/news/zenphoto-admin-chosen-dropdowns/). Since you provide chosen.jquery.min.js with bulk-move, there is no need to point to any depending plugin.

Thank you for your bulk-move plugin!